💡 chore    : external_id 필드 추가 및 테스트 수정

+++ apps/lecture/models/lecture.py
@@ -33,7 +33,7 @@ class CrawledLecture(UUIDBaseModel):
     url_link = models.CharField(max_length=500, null=False)
     thumbnail_img_url = models.CharField(max_length=500, null=True, blank=True)
 
-    external_id = models.IntegerField()
+    external_id = models.BigIntegerField()
 
     categories: "ManyToManyField[Category,Category]" = models.ManyToManyField(
         "Category", through="LectureCategory", related_name="crawled_lecture_categories"
